Correct Height()
Refactored height() such that it returns the maximum height of a node.
This is acheived by getting the level of each node. This is acheived by
incrementing a counter every time a child node is accessed. Once a leaf
node is reached, it is pushed to a queue. Recursion is used to access
child nodes and reassign the counter and queue.