Create IsBalanced()
This recursively traverses the tree, and checks each node if it has only
one branch with a height of 2 or greater. If so, push the false value to
an array. At the end of the function, check if false is in the array.
If so, return false. If not, return true.